2005-07-19  Willem Jan Palenstijn <wjpalenstijn at users.sourceforge.net> 	* views/DollWidget.cpp: fix undefined return value causing crashes 	* various: fixed warnings 	* configure.ac: enable -Wall by default

+++ sound/SoundManager.cpp
@@ -581,7 +581,7 @@ void SoundManager::update()
       m_pCurrentSong = SoundManager::RequestSong (m_CurrentGroup);
       if(m_pCurrentSong)
         {
-          printf ("assigning new song! %x\n", m_pCurrentSong);
+          printf ("assigning new song! %x\n", (unsigned int)m_pCurrentSong);
           if(!m_pCurrentSong->Play (false))
             {
               printf ("play failed!\n");
